Refactored functions into helper classes
- Refactored download function into DownloadHelper.swift
- Refactored document function into Document Helper.swift
