+++ akarsh seggemu resume/Controller/MainTableViewController.swift
@@ -54,18 +54,11 @@ class MainTableViewController: UITableViewController {
         target.labelContentResumeSchemaTableViewHeader = mainCell.nameLabelOfLanguage.text
     }
     
-    // return documents directory as output
-    func getDocumentsDirectory() -> URL {
-        let paths = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)
-        let documentsDirectory = paths[0]
-        return documentsDirectory
-    }
-    
     // MARK: - Fetch Data to create the JSON file in the documents directory
     
     func fetchData(_ resumeFileName: String, _ resumeURL: String) {
         // Create destination URL
-        let documentsUrl: URL = getDocumentsDirectory()
+        let documentsUrl: URL = DocumentHelper.getDocumentsDirectory()
         // Get the file path in documents directory
         let destinationFileUrl = documentsUrl.appendingPathComponent(resumeFileName)
         
@@ -91,36 +84,7 @@ class MainTableViewController: UITableViewController {
         }
         
         // url to the resume JSON file
-        downloadFileFromUrl(englishResumeURL, destinationFileUrl)
-    }
-    
-    // Download file from the URL and store it in the destination File URL
-    fileprivate func downloadFileFromUrl(_ url: String, _ destinationFileUrl: URL) {
-        if let urlString = URL(string: url) {
-            let sessionConfig = URLSession(configuration: .default)
-            let request = URLRequest(url: urlString)
-            // download task to download the resume JSON file
-            let dataTask = sessionConfig.downloadTask(with: request) { data, response, error in
-                if let tempLocalUrl = data, error == nil {
-                    // if success print the status code of 200
-                    if let statusCode = (response as? HTTPURLResponse)?.statusCode {
-                        print("Successfully downloaded. Status code: \(statusCode)")
-                    }
-                    
-                    do {
-                        // copying the file to the destination file path
-                        try FileManager.default.copyItem(at: tempLocalUrl, to: destinationFileUrl)
-                        print("File created at \(destinationFileUrl)")
-                    } catch let writeError {
-                        print("Error creating a file \(destinationFileUrl) : \(writeError)")
-                    }
-                    
-                } else {
-                    print("Error took place while downloading a file. Error description: %@ \(String(describing: error?.localizedDescription))")
-                }
-            }
-            dataTask.resume()
-        }
+        DownloadHelper.extractedFunc(resumeURL, destinationFileUrl)
     }
 }
