+++ akarsh seggemu resume/DocumentHelper.swift
@@ -1,0 +1,18 @@
+//
+//  DocumentHelper.swift
+//  akarsh seggemu resume
+//
+//  Created by Akarsh Seggemu on 24/01/2019.
+//  Copyright © 2019 Akarsh Seggemu. All rights reserved.
+//
+
+import Foundation
+
+class DocumentHelper {
+    // return documents directory as output
+    static func getDocumentsDirectory() -> URL {
+        let paths = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)
+        let documentsDirectory = paths[0]
+        return documentsDirectory
+    }
+}
